backend/requirements.txt
fastapi[standard]
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.6.0